# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API spesifikasi untuk Aplikasi Persuratan Internal
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  message: string;
  error?: string;
}

export interface SuccessResponse {
  message: string;
}

export interface LoginRequest {
  /** Username pengguna */
  username: string;
  /** Password pengguna */
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  staff: "staff",
  manager: "manager",
} as const;

export interface User {
  id: number;
  username: string;
  namaLengkap: string;
  role: UserRole;
  email: string;
}

export interface LoginResponse {
  /** JWT access token */
  token: string;
  user: User;
}

/**
 * Jenis surat (masuk atau keluar)
 */
export type CreateSuratRequestJenis =
  (typeof CreateSuratRequestJenis)[keyof typeof CreateSuratRequestJenis];

export const CreateSuratRequestJenis = {
  masuk: "masuk",
  keluar: "keluar",
} as const;

export interface CreateSuratRequest {
  /** Nomor surat resmi */
  nomorSurat: string;
  /** Perihal / subjek surat */
  perihal: string;
  /** Nama pengirim surat */
  pengirim: string;
  /** Nama penerima surat */
  penerima: string;
  /** Jenis surat (masuk atau keluar) */
  jenis: CreateSuratRequestJenis;
  /** Tanggal surat (ISO date string) */
  tanggalSurat: string;
  /** Keterangan tambahan (opsional) */
  keterangan?: string;
  /** File surat (PDF atau gambar) */
  file?: Blob;
}

export type SuratJenis = (typeof SuratJenis)[keyof typeof SuratJenis];

export const SuratJenis = {
  masuk: "masuk",
  keluar: "keluar",
} as const;

export interface Surat {
  id: number;
  nomorSurat: string;
  perihal: string;
  pengirim: string;
  penerima: string;
  jenis: SuratJenis;
  tanggalSurat: string;
  keterangan?: string | null;
  namaFile?: string | null;
  urlFile?: string | null;
  createdById: number;
  createdByName: string;
  createdAt: string;
}

export interface SuratListResponse {
  data: Surat[];
  total: number;
  page: number;
  limit: number;
}

export interface DashboardStats {
  /** Total semua surat */
  totalSurat: number;
  /** Total surat masuk */
  suratMasuk: number;
  /** Total surat keluar */
  suratKeluar: number;
  /** Surat yang masuk/keluar bulan ini */
  suratBulanIni: number;
  /** Surat yang masuk/keluar minggu ini */
  suratMingguIni: number;
}

export type GetSuratListParams = {
  /**
   * Filter berdasarkan jenis surat
   */
  jenis?: GetSuratListJenis;
  /**
   * Nomor halaman
   * @minimum 1
   */
  page?: number;
  /**
   * Jumlah item per halaman
   * @minimum 1
   * @maximum 100
   */
  limit?: number;
};

export type GetSuratListJenis =
  (typeof GetSuratListJenis)[keyof typeof GetSuratListJenis];

export const GetSuratListJenis = {
  masuk: "masuk",
  keluar: "keluar",
} as const;