# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API spesifikasi untuk Aplikasi Persuratan Internal
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Login pengguna
 */
export const LoginBody = zod.object({
  username: zod.string().describe("Username pengguna"),
  password: zod.string().describe("Password pengguna"),
});

export const LoginResponse = zod.object({
  token: zod.string().describe("JWT access token"),
  user: zod.object({
    id: zod.number(),
    username: zod.string(),
    namaLengkap: zod.string(),
    role: zod.enum(["admin", "staff", "manager"]),
    email: zod.string(),
  }),
});

/**
 * @summary Ambil profil pengguna saat ini
 */
export const GetMeResponse = zod.object({
  id: zod.number(),
  username: zod.string(),
  namaLengkap: zod.string(),
  role: zod.enum(["admin", "staff", "manager"]),
  email: zod.string(),
});

/**
 * @summary Ambil daftar semua surat
 */

export const getSuratListQueryLimitMax = 100;

export const GetSuratListQueryParams = zod.object({
  jenis: zod
    .enum(["masuk", "keluar"])
    .optional()
    .describe("Filter berdasarkan jenis surat"),
  page: zod.coerce.number().min(1).optional().describe("Nomor halaman"),
  limit: zod.coerce
    .number()
    .min(1)
    .max(getSuratListQueryLimitMax)
    .optional()
    .describe("Jumlah item per halaman"),
});

export const GetSuratListResponse = zod.object({
  data: zod.array(
    zod.object({
      id: zod.number(),
      nomorSurat: zod.string(),
      perihal: zod.string(),
      pengirim: zod.string(),
      penerima: zod.string(),
      jenis: zod.enum(["masuk", "keluar"]),
      tanggalSurat: zod.string(),
      keterangan: zod.string().nullish(),
      namaFile: zod.string().nullish(),
      urlFile: zod.string().nullish(),
      createdById: zod.number(),
      createdByName: zod.string(),
      createdAt: zod.string(),
    }),
  ),
  total: zod.number(),
  page: zod.number(),
  limit: zod.number(),
});

/**
 * @summary Buat surat baru dengan upload file
 */
export const CreateSuratBody = zod.object({
  nomorSurat: zod.string().describe("Nomor surat resmi"),
  perihal: zod.string().describe("Perihal \/ subjek surat"),
  pengirim: zod.string().describe("Nama pengirim surat"),
  penerima: zod.string().describe("Nama penerima surat"),
  jenis: zod
    .enum(["masuk", "keluar"])
    .describe("Jenis surat (masuk atau keluar)"),
  tanggalSurat: zod.string().describe("Tanggal surat (ISO date string)"),
  keterangan: zod
    .string()
    .optional()
    .describe("Keterangan tambahan (opsional)"),
  file: zod
    .instanceof(File)
    .optional()
    .describe("File surat (PDF atau gambar)"),
});

/**
 * @summary Ambil detail surat berdasarkan ID
 */
export const GetSuratByIdParams = zod.object({
  id: zod.coerce.number(),
});

export const GetSuratByIdResponse = zod.object({
  id: zod.number(),
  nomorSurat: zod.string(),
  perihal: zod.string(),
  pengirim: zod.string(),
  penerima: zod.string(),
  jenis: zod.enum(["masuk", "keluar"]),
  tanggalSurat: zod.string(),
  keterangan: zod.string().nullish(),
  namaFile: zod.string().nullish(),
  urlFile: zod.string().nullish(),
  createdById: zod.number(),
  createdByName: zod.string(),
  createdAt: zod.string(),
});

/**
 * @summary Hapus surat berdasarkan ID
 */
export const DeleteSuratParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteSuratResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary Ambil statistik ringkasan untuk dashboard
 */
export const GetDashboardStatsResponse = zod.object({
  totalSurat: zod.number().describe("Total semua surat"),
  suratMasuk: zod.number().describe("Total surat masuk"),
  suratKeluar: zod.number().describe("Total surat keluar"),
  suratBulanIni: zod.number().describe("Surat yang masuk\/keluar bulan ini"),
  suratMingguIni: zod.number().describe("Surat yang masuk\/keluar minggu ini"),
});

/**
 * @summary Ambil surat-surat terbaru untuk aktivitas feed
 */
export const GetRecentSuratResponseItem = zod.object({
  id: zod.number(),
  nomorSurat: zod.string(),
  perihal: zod.string(),
  pengirim: zod.string(),
  penerima: zod.string(),
  jenis: zod.enum(["masuk", "keluar"]),
  tanggalSurat: zod.string(),
  keterangan: zod.string().nullish(),
  namaFile: zod.string().nullish(),
  urlFile: zod.string().nullish(),
  createdById: zod.number(),
  createdByName: zod.string(),
  createdAt: zod.string(),
});
export const GetRecentSuratResponse = zod.array(GetRecentSuratResponseItem);