# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API spesifikasi untuk Aplikasi Persuratan Internal
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardStats {
  /** Total semua surat */
  totalSurat: number;
  /** Total surat masuk */
  suratMasuk: number;
  /** Total surat keluar */
  suratKeluar: number;
  /** Surat yang masuk/keluar bulan ini */
  suratBulanIni: number;
  /** Surat yang masuk/keluar minggu ini */
  suratMingguIni: number;
}