# lib/api-zod/src/generated/types/surat.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API spesifikasi untuk Aplikasi Persuratan Internal
 * OpenAPI spec version: 0.1.0
 */
import type { SuratJenis } from "./suratJenis";

export interface Surat {
  id: number;
  nomorSurat: string;
  perihal: string;
  pengirim: string;
  penerima: string;
  jenis: SuratJenis;
  tanggalSurat: string;
  keterangan?: string | null;
  namaFile?: string | null;
  urlFile?: string | null;
  createdById: number;
  createdByName: string;
  createdAt: string;
}