# lib/api-zod/src/generated/types/suratListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API spesifikasi untuk Aplikasi Persuratan Internal
 * OpenAPI spec version: 0.1.0
 */
import type { Surat } from "./surat";

export interface SuratListResponse {
  data: Surat[];
  total: number;
  page: number;
  limit: number;
}